backend/requirements.txt
yt-dlp
pyspellchecker
django
django-cors-headers
ffmpeg-python
opencv-python-headless
Pillow
openai
gunicorn

## 接下来就是pytorch和faster_whisper的部分了##
# sudo apt update
# sudo apt install -y ffmpeg # 系统的ffmpeg